backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.1
redis==5.0.1
celery==5.3.4
requests==2.31.0
pypdf==4.0.1
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
markdown==3.5.2
bleach==6.1.0
numpy==1.26.4
scikit-learn==1.4.0
openai==1.12.0
python-dotenv==1.0.1
